fix(plugins): keep the plugin query tests off the plugin registry
The read-only query API added in #1784 is exercised through
`createAppAPI`, which lives in `usePlugins.ts`. That module imports the
whole built-in plugin registry, so loading it in a Node test pulled in
MapCanvas, CesiumCanvas, and every `maplibre-*` plugin: 39 browser-only
modules, none of them meaningfully exercised, all of them newly counted
by the coverage reporter.

Coverage is reported only over files a test actually imports, so those 39
files landed in the denominator at 1-30% function coverage and dropped
the total from 72.90% to 60.36%, under the 63% floor. CI has been red on
`main` since, and the report is measuring module reachability rather than
how well the code is tested.

Move the six query methods and `readPluginSelection` into
`lib/plugin-layer-queries.ts`, which needs only the store, and have
`createAppAPI` spread them in. The test imports that module directly and
no longer stubs `maplibre-gl`, `window`, `sessionStorage`, or
`localStorage` to get off the ground. Same reasoning, and the same shape,
as `geo-editor-geometry.ts` in `@geolibre/plugins`, which is already kept
free of the Geoman/MapLibre runtime so it can be unit-tested under Node.

`SKETCHES_SOURCE_KIND` gets a subpath export alongside the five already
in that package, so the test can share the constant instead of repeating
the string literal as it did before.

The wiring into the plugin-facing API is now a typed spread rather than
something this test asserts; `tsc` covers it on every build.

Coverage returns to 82.84% lines / 84.45% branches / 72.55% functions
(baseline before #1784: 82.89 / 84.46 / 72.90), with the counted file set
back from 444 to 406. The one addition is the extracted module itself, at
100% lines and 95.45% functions. All 8 tests still pass.
