+++ packages/plugins/package.json
@@ -8,6 +8,7 @@
   "exports": {
     ".": "./src/index.ts",
     "./cog-spectral-profile": "./src/plugins/cog-spectral-profile.ts",
+    "./geo-editor-geometry": "./src/plugins/geo-editor-geometry.ts",
     "./local-netcdf": "./src/plugins/local-netcdf.ts",
     "./maplibre-graticule": "./src/plugins/maplibre-graticule.ts",
     "./raster-symbology": "./src/plugins/raster-symbology.ts",
